requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
azure-core==1.26.4
azure-storage-blob==12.16.0
billiard==3.6.4.0
black==24.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
Django==4.1.5
django-allauth==0.52.0
django-colorfield==0.8.0
django-jazzmin==2.6.0
django-multiupload==0.6.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
isodate==0.6.1
kombu==5.2.4
mypy-extensions==1.0.0
numpy==1.24.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pandas==1.5.3
pathspec==0.12.1
platformdirs==4.2.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycountry==22.3.5
pycparser==2.21
PyJWT==2.6.0
pypng==0.20220715.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
qrcode==7.4.2
redis==4.5.4
requests-oauthlib==1.3.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
cryptography==42.0.5
msal==1.28.0
pycparser==2.21
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
Pillow==9.4.0
django-simple-history